+++ src/components/common/Input/Input.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Input from '../Input';
+import Input from './Input';
 import { useState } from 'react';
 
 const meta: Meta<typeof Input> = {
